requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
click==8.1.8
eclipse-zenoh==1.3.0
fastapi==0.115.12
h11==0.14.0
idna==3.10
numpy==2.2.4
pi-plates==10.5
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pytz==2025.2
RPi.bme280==0.2.4
RPi.GPIO==0.7.1
six==1.17.0
smbus2==0.5.0
sniffio==1.3.1
spidev==3.6
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
